Remove redundant cast

+++ core/trino-spi/src/main/java/io/trino/spi/block/Int128ArrayBlock.java
@@ -118,7 +118,7 @@ public class Int128ArrayBlock
         if (valueIsNull != null) {
             consumer.accept(valueIsNull, sizeOf(valueIsNull));
         }
-        consumer.accept(this, (long) INSTANCE_SIZE);
+        consumer.accept(this, INSTANCE_SIZE);
     }
 
     @Override
